feat(m11-s7): cpus 1-decimal precision + drop "0" sentinel in describe
Closes todo #67 (cpus precision normalisation at the HTTP boundary)
and todo #75 (sandbox describe / inspect rendering "CPUs: 0,
Memory: 0 MB" for default-resource container sessions).

Wire-shape change: BackendSpecific::Container { cpus } widens from
u32 to f32 so the spec's 1-decimal precision (0.8, 1.5, …) is
representable end-to-end. CreateSessionRequest.cpus and
SessionConfigDto.cpus widen the same way. Lima's
BackendSpecific::Lima { cpus: u32 } is unchanged — Lima's spec'd
values are integer cores via QEMU -smp.

Persistence: SessionConfig keeps cpus: u32 for rollback safety and
adds a sibling cpus_decimal: Option<f32> with #[serde(default)] +
skip_serializing_if. When a fractional value is requested both fields
are populated (the integer field takes the floor); the DTO mapper and
runtime projector prefer cpus_decimal when present. Older daemons
rolling back still see a usable integer per CLAUDE.md's persistence
rules.

Boundary normalisation: a new round_cpus_one_decimal helper rounds
incoming values via (f * 10).round() / 10 so 0.81 → 0.8, mirroring
compute_default_resource_limits's existing rounding. The lossy `as
f64` widen in resource_ceilings is gone; the truncation that turned
1.5 into 1 is fixed.

Render polish: sandbox describe / inspect for a container session
created without --cpus / --memory now renders the resolved host-80%
default with a "(default)" hint, e.g. "CPUs: 1.6 (default)" instead
of "CPUs: 0". Integer-valued cpus (Lima default 2, explicit container
4) render without trailing ".0" for visual continuity with pre-S7
output.

+++ sandboxd/sandbox-core/src/api/dto.rs
@@ -186,7 +186,15 @@ pub struct SessionMountInfo {
 ///    enum object, which is what CLI consumers actually want to print.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SessionConfigDto {
-    pub cpus: u32,
+    /// User-requested CPU ceiling. M11-S7 todo #67 widened this from
+    /// `u32` to `f32` so the spec § "Resource defaults — container
+    /// only" 1-decimal precision survives the persisted round-trip.
+    /// Lima sessions still see whole-number values; container
+    /// sessions see whatever the operator passed (e.g. `1.5`) or
+    /// `0.0` when the operator omitted the flag and the daemon will
+    /// substitute the host-80% default at runtime — the resolved
+    /// applied value lives in [`Self::resolved_cpus`].
+    pub cpus: f32,
     pub memory_mb: u32,
     pub disk_gb: u32,
     /// Effective CPU ceiling actually applied at runtime.
